requirements.txt
amqp==2.1.4
appdirs==1.4.2
billiard==3.5.0.2
celery==4.0.2
cffi==1.9.1
cryptography==1.7.2
decorator==4.0.11
Django==1.10.5
django-celery-results==1.0.1
django-countries==4.1
django-filter==1.0.1
django-simple-history==1.8.2
djangorestframework==3.5.4
idna==2.2
ipython==5.2.2
ipython-genutils==0.1.0
kombu==4.0.2
Markdown==2.6.8
packaging==16.8
paramiko==2.1.2
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
pyasn1==0.2.2
pycparser==2.17
Pygments==2.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
redis==2.10.5
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
vine==1.1.3
wcwidth==0.1.7
wheel==0.24.0
